Finalize API refactor
