+++ app/api_v1/errors/__init__.py
@@ -1,1 +1,1 @@
-from app.api_v1.errors import errors, fhir_errors, exceptions
+from app.api_v1.errors import errors, fhir, exceptions
